Feature #11 Closed.  add_attraction is updated to parse the sections into a array which is stored in the format [[title,depth],[title2,depth2]].  Only the attraction Operation has sections so when working with PasswordGeneration don't expect an outline.
Had to save the outline as a string which is evaled when it is retrieved, and parsed when it is set.  This is due to to the fact that ActiveRecord has no Object template for storing items in the database.  Have not used ActiveRecord in a while so will do some searching later for a better way.